pgkernelgateway/src/test/javascript/spec/app/entities/user-connection/user-connection.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { UserConnectionService } from 'app/entities/user-connection/user-connection.service';
import { IUserConnection, UserConnection } from 'app/shared/model/user-connection.model';

describe('Service Tests', () => {
  describe('UserConnection Service', () => {
    let injector: TestBed;
    let service: UserConnectionService;
    let httpMock: HttpTestingController;
    let elemDefault: IUserConnection;
    let expectedResult;
    let currentDate: moment.Moment;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = {};
      injector = getTestBed();
      service = injector.get(UserConnectionService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new UserConnection(0, currentDate, currentDate);
    });

    describe('Service methods', () => {
      it('should find an element', async () => {
        const returnedFromService = Object.assign(
          {
            loginDate: currentDate.format(DATE_TIME_FORMAT),
            logoutDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: elemDefault });
      });

      it('should create a UserConnection', async () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            loginDate: currentDate.format(DATE_TIME_FORMAT),
            logoutDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );
        const expected = Object.assign(
          {
            loginDate: currentDate,
            logoutDate: currentDate
          },
          returnedFromService
        );
        service
          .create(new UserConnection(null))
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should update a UserConnection', async () => {
        const returnedFromService = Object.assign(
          {
            loginDate: currentDate.format(DATE_TIME_FORMAT),
            logoutDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            loginDate: currentDate,
            logoutDate: currentDate
          },
          returnedFromService
        );
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should return a list of UserConnection', async () => {
        const returnedFromService = Object.assign(
          {
            loginDate: currentDate.format(DATE_TIME_FORMAT),
            logoutDate: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );
        const expected = Object.assign(
          {
            loginDate: currentDate,
            logoutDate: currentDate
          },
          returnedFromService
        );
        service
          .query(expected)
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a UserConnection', async () => {
        const rxPromise = service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
